requirements.txt
Django==4.1.6
djangorestframework==3.14.0
Pygments==2.11.2
psycopg2>=2.8

